backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
nibabel>=5.2,<6.0
numpy>=1.26,<3.0
scipy>=1.13,<2.0
pydantic>=2.7,<3.0
Pillow>=10.3,<12.0
